Implement environment directory extension to accommodate #28.
The new system allows administrators to still control the storage of environment files, although now by supplying additional methods on a common function rather than merely controlling a special variable for the root. The new system also allows further kinds of files, such as generic runtime data and template overrides.

Unfortunately since the locations need to be changed in order to implement useful default locations for the newly distributed systems, we also end up breaking existing setups alongside the backwards compatibility that we lose due to the removal of *environment-root*.

[Compatibility Breaking] [Setup Breaking]